# feat: improve input API (BREAKING)