--- a/src/fingerprint.ts
+++ b/src/fingerprint.ts
@@ -1,15 +1,9 @@
 import pLimit from "p-limit";
 
 import { DEFAULT_CONCURRENCY } from "./constants.js";
-import { calculateContentHash } from "./inputs/content.js";
+import { calculateContentHashes } from "./inputs/content.js";
 import { calculateFileHash, calculateFileHashSync } from "./inputs/file.js";
-import type {
-  Config,
-  ContentInput,
-  ContentValue,
-  Fingerprint,
-  FingerprintOptions,
-} from "./types.js";
+import type { Config, Fingerprint, FingerprintOptions } from "./types.js";
 import { getInputFiles, getInputFilesSync, mergeHashes } from "./utils.js";
 
 export async function calculateFingerprint(
@@ -32,9 +26,7 @@ export async function calculateFingerprint(
     inputFiles.map((path) => limit(() => calculateFileHash(path, config))),
   );
 
-  const contentHashes = content
-    ? getContentInputs(content)?.map((input) => calculateContentHash(input, config))
-    : [];
+  const contentHashes = content ? calculateContentHashes(content, config) : [];
 
   return mergeHashes(fileHashes, contentHashes, config);
 }
@@ -55,16 +47,7 @@ export function calculateFingerprintSync(
   });
 
   const fileHashes = inputFiles.map((path) => calculateFileHashSync(path, config));
-
-  const contentHashes = content
-    ? getContentInputs(content)?.map((input) => calculateContentHash(input, config))
-    : [];
+  const contentHashes = content ? calculateContentHashes(content, config) : [];
 
   return mergeHashes(fileHashes, contentHashes, config);
-}
-
-function getContentInputs(content: Record<string, ContentValue>): ContentInput[] {
-  return Object.entries(content).map(([key, value]) => {
-    return { key, ...value };
-  });
 }