--- a/src/fingerprint.ts
+++ b/src/fingerprint.ts
@@ -1,9 +1,9 @@
 import pLimit from "p-limit";
 
 import { DEFAULT_CONCURRENCY } from "./constants.js";
-import { calculateContentHashes } from "./inputs/content.js";
+import { calculateContentHash } from "./inputs/content.js";
 import { calculateFileHash, calculateFileHashSync } from "./inputs/file.js";
-import type { Config, Fingerprint, FingerprintOptions } from "./types.js";
+import type { Config, ContentHash, Fingerprint, FingerprintOptions, InputRecord } from "./types.js";
 import { getInputFiles, getInputFilesSync, mergeHashes } from "./utils.js";
 
 export async function calculateFingerprint(
@@ -21,7 +21,7 @@ export async function calculateFingerprint(
     inputFiles.map((path) => limit(() => calculateFileHash(path, config))),
   );
 
-  const contentHashes = content ? calculateContentHashes(content, config) : [];
+  const contentHashes = calculateContentHashes(content, config);
   return mergeHashes(fileHashes, contentHashes, config);
 }
 
@@ -37,6 +37,14 @@ export function calculateFingerprintSync(
   const inputFiles = getInputFilesSync(basePath, { files, ignores });
   const fileHashes = inputFiles.map((path) => calculateFileHashSync(path, config));
 
-  const contentHashes = content ? calculateContentHashes(content, config) : [];
+  const contentHashes = calculateContentHashes(content, config);
   return mergeHashes(fileHashes, contentHashes, config);
+}
+
+function calculateContentHashes(inputs: InputRecord | undefined, config: Config): ContentHash[] {
+  if (!inputs) return [];
+
+  return Object.entries(inputs).map(([key, value]) =>
+    calculateContentHash({ key, ...value }, config),
+  );
 }