--- a/src/fingerprint.ts
+++ b/src/fingerprint.ts
@@ -3,12 +3,12 @@ import pLimit from "p-limit";
 import { DEFAULT_CONCURRENCY } from "./constants.js";
 import { calculateContentHash } from "./inputs/content.js";
 import { calculateFileHash, calculateFileHashSync } from "./inputs/file.js";
-import type { Config, ContentHash, Fingerprint, FingerprintOptions, InputRecord } from "./types.js";
+import type { Config, Fingerprint, FingerprintOptions } from "./types.js";
 import { getInputFiles, getInputFilesSync, mergeHashes } from "./utils.js";
 
 export async function calculateFingerprint(
   basePath: string,
-  { hashAlgorithm, files, ignores, extraInputs: content, concurrency }: FingerprintOptions = {},
+  { hashAlgorithm, files, ignores, contentInputs: content, concurrency }: FingerprintOptions = {},
 ): Promise<Fingerprint> {
   const config: Config = {
     basePath,
@@ -21,13 +21,13 @@ export async function calculateFingerprint(
     inputFiles.map((path) => limit(() => calculateFileHash(path, config))),
   );
 
-  const contentHashes = calculateContentHashes(content, config);
+  const contentHashes = content?.map((input) => calculateContentHash(input, config)) ?? [];
   return mergeHashes(fileHashes, contentHashes, config);
 }
 
 export function calculateFingerprintSync(
   basePath: string,
-  { hashAlgorithm, files, ignores, extraInputs: content }: FingerprintOptions = {},
+  { hashAlgorithm, files, ignores, contentInputs: content }: FingerprintOptions = {},
 ): Fingerprint {
   const config: Config = {
     basePath,
@@ -36,15 +36,6 @@ export function calculateFingerprintSync(
 
   const inputFiles = getInputFilesSync(basePath, { files, ignores });
   const fileHashes = inputFiles.map((path) => calculateFileHashSync(path, config));
-
-  const contentHashes = calculateContentHashes(content, config);
+  const contentHashes = content?.map((input) => calculateContentHash(input, config)) ?? [];
   return mergeHashes(fileHashes, contentHashes, config);
-}
-
-function calculateContentHashes(inputs: InputRecord | undefined, config: Config): ContentHash[] {
-  if (!inputs) return [];
-
-  return Object.entries(inputs).map(([key, value]) =>
-    calculateContentHash({ key, ...value }, config),
-  );
 }