… .github/workflows/pkg-size.yml
… README.md
… benchmarks/fingerprint.bench.ts
… src/__tests__/fingerprint.test.ts
… src/__tests__/git.test.ts
… src/__tests__/utils.test.ts
… src/index.ts
… src/inputs/__tests__/content.test.ts
… src/inputs/__tests__/file.test.ts
… src/inputs/content.ts
… src/inputs/file.ts
… src/types.ts
… src/utils.ts
… test-utils/format.ts
… test-utils/fs.ts