## What does this PR do?

Based on the experience so far, improve input API. This is a breaking change.

## Relevant implementation details

Use ESLint option names instead of TSC options:
- `rootDir` => `basePath`
- `include` => `files`
- `exclude` => `ignores`
- `extraInputs` => `contentInputs`

## How did you test this?

All tests should pass. 

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - Core API and option keys renamed for consistency (rootDir → basePath; include → files; exclude → ignores; extraInputs → contentInputs); several function/return names updated (e.g., remapPaths → rebasePath, hashContent → hashData).
- New Features
  - Added content helpers (textContent, jsonContent, envContent), exported content-hashing helpers, new sortBy and test-format helper.
- Documentation
  - README updated to reflect new option names, concurrency option, examples, and API signatures.
- Chores
  - CI workflow renamed from “Check Package Size” to “Package Size”.
- Tests
  - Test suites and utilities updated to use basePath and renamed options.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->