@mdjastrzebski 2025-09-27
📊 Performance Comparison (vs baseline):
| Task name                     | Baseline latency (ms) | Current latency (ms) | Change (ms)    |
| ----------------------------- | --------------------- | -------------------- | -------------- |
| lodash-sync                   | 16.02                 | 15.60                | -0.41 (-2.6%)  |
| lodash                        | 18.13                 | 18.30                | +0.17 (+0.9%)  |
| express-sync                  | 7.85                  | 7.68                 | -0.17 (-2.2%)  |
| express                       | 7.95                  | 8.14                 | +0.19 (+2.4%)  |
| react-sync                    | 314.12                | 304.87               | -9.25 (-2.9%)  |
| react                         | 357.52                | 354.56               | -2.95 (-0.8%)  |
| react-native-sync             | 512.51                | 474.64               | -37.88 (-7.4%) |
| react-native                  | 551.87                | 508.26               | -43.62 (-7.9%) |
| react-native (null hash)-sync | 68.25                 | 62.52                | -5.72 (-8.4%)  |
| react-native (null hash)      | 23.39                 | 22.77                | -0.62 (-2.6%)  |
| expensify-ios-sync            | 7.55                  | 7.58                 | +0.03 (+0.4%)  |
| expensify-ios                 | 8.03                  | 8.17                 | +0.14 (+1.7%)  |
| expensify-android-sync        | 16.42                 | 16.26                | -0.16 (-1.0%)  |
| expensify-android             | 16.95                 | 16.81                | -0.15 (-0.9%)  |